Fix loading session files from geany.conf at startup.
This regression was introduced by 343bc1869e5d669d5d652bef7ec735d5db7ecd67,
which introduces a fallback to [relative_files] if a file from the [files]
section was not found. This check was buggy since it used the unescaped
version of the file name (that has %2F instead of /) which makes the
check always fail.