Add new GeanyPage class.
The purpose of this class is to encapsulate notebook tabs and their contents,
as well as active page handling. This includes management of the tab label and
icon. As such, GeanyPage separates GeanyDocument and our GtkNotebook logic, so
that notebook.c doesn't know it displays GeanyDocuments/Scintillas and
document.c isn't concerned with how to draw the current document.