notebooks: Update MRU list when the focus changes in the document notebooks
Previously it was only updated when switching tabs. This leads to the MRU
not being updated when notebooks are switched (without changing their respective
current page). This commit fixes it.

This commit also enforces that, when clicking the tabs itself on the notebooks,
that the corresponding sci widget is focused.