scimethods: Generate ScintillaObject methods from sciwrappers.h
These methods are the OOP counterpart of the sci_* functions. They simply have
a different name so that g-ir-scanne picks them up as instance methods.

Because they just have a different name and are otherwise identical to sci_*
they can be generated using a small script. Checked-in to git to avoid build
dependancy on python.