Introduce GEANY_EXPORT as G_MODULE_EXPORT wrapper
The wrapper is because G_MODULE_EXPORT can't be used if/when we want to
switch to hidden-by-default visibility and only export explicit functions
because G_MODULE_EXPORT is always empty on Linux.

Change glade callbacks to use GEANY_EXPORT. Also remove it from the declaraitons
so that it's only at the definition.