plugins: Use libpeas[1] to manage plugins
libpeas provides a uniform API to manage plugins (extensions) implemented
in multiple languages. We can offload lots of the internal management, loading
and metadata code to it.

Apart from offloading lots of our custom plugin code, we should be able
to exploit libpeas to support python (and other languages) plugins out of the
box in the future.

The major changes for plugins:
1) Metadata is not part of the plugin executable anymore. Instead, a keyfile-style
.plugin file provides all the necessary, localized, metadata.
2) Instantiation of a plugin involves a GObject that implements a given
GInterface. We define this GInterface, and provide an abstract base class to
help plugins do that.

Other than that, Geany's APIs are accessed normally. Therefore only bootstrapping
of the plugins is affected.

To relieve plugin developers from being required to make immediate changes
to their plugin code we ship a compatibility C source file that transparently
bridges existing plugins to the new system. Just compile and link this file
and the plugin will continue to work normally.

[1]: https://developer.gnome.org/libpeas/1.12/