keybindings: Export keybindings_set_group_with_handler() to plugin api
Plugins can use this instead of keybindings_set_group_with_handler(). This
provides no automatic freeing when the plugin is unloaded and has to be done
manually.

This is intended mainly for the GeanyGI gobject api wrapper that does
keybindings_free_group() in its destructor, and has no reference to
the plugin instance.