utils: add functions to process file name list
1) utils_strv_find_common_prefix
Locates the common prefix.

2) utils_strv_find_lcs
Finds the longest common substring.

3) utils_strv_shorten_file_list
Transforms the file list by removing the common prefix and ellipsizing
the longest common substring. This is intended to be used for fixing #1069.

Although only 3 will be used immediately, I separated the functionality, so
that the other two function can be used on their own.