sidebar: refactor get_doc_parent()
It checked documents_show_paths whether there is anything to do.
Now the caller is responsible for that. Also, instead of returning
a static iter, the caller passes an iter which is modified accordingly
(out parameter). The previous out parameter (new_parent) became the
return value. This seems more conventional coding style.