sidebar: correct and simplify check against splitting home dir
The intent is to have dirs below the home directory of the user in a separate
root node (i.e. even if other files below /home are open).

The old check was hard to follow and apparently wrong because it didn't work
in all cases (in some situations there was a node labeled '~' as a child
node of /home).

Unlike the distinct tree node for projects, this does not prevent merging paths,
so if only files below /home/user/dev are open, then the tree node should be
labeled "~/dev".