sidebar: simplify code by handling projects and home directory equally
With the sidebar tree, the project docs were handled differently to docs inside
the home dir. The project handling is better because it does not need the
logic to not split parent paths leading to the base directory. The code
handling the home directory had extra logic but since we basically want to
treat projects and the home directory the same, unifiy the code using the
simpler one.

Also get rid of unused members in TreeForeachData.